+++ crates/jsonschema-py/python/jsonschema_rs/__init__.py
@@ -46,6 +46,7 @@ class ValidationError(ValueError):
     evaluation_path: list[str | int]
     kind: ValidationErrorKind
     instance: Any
+    absolute_keyword_location: str | None
 
     def __init__(
         self,
@@ -56,6 +57,7 @@ class ValidationError(ValueError):
         evaluation_path: list[str | int],
         kind: ValidationErrorKind,
         instance: Any,
+        absolute_keyword_location: str | None = None,
     ) -> None:
         super().__init__(verbose_message)
         self.message = message
@@ -65,6 +67,7 @@ class ValidationError(ValueError):
         self.evaluation_path = evaluation_path
         self.kind = kind
         self.instance = instance
+        self.absolute_keyword_location = absolute_keyword_location
 
     def __str__(self) -> str:
         return self.verbose_message